refactor: calculateLottoPrize관련 리팩토링
