+++ docs/리뷰개선사항들.md
@@ -55,7 +55,7 @@
 
 ### calculate 관련 개선사항
 
-- [ ] lastLottoNumberList와 lastLottoNumbers간의 차이를 알기 힘듬 네이밍 변경하기
-- [ ] calculateLottoPrize에서 filter callback boolean리턴해주기
-- [ ] switch문에서 바로 값 return하기
-- [ ] 등수 체크하는 로직이 updateObj 로직과 겹침 분리
+- [x] lastLottoNumberList와 lastLottoNumbers간의 차이를 알기 힘듬 네이밍 변경하기
+- [x] calculateLottoPrize에서 filter callback boolean리턴해주기
+- [x] switch문에서 바로 값 return하기
+- [x] 등수 체크하는 로직이 updateObj 로직과 겹침 분리
